# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Dokumentasi Produk",
        "description": "Dokumentasi API untuk pengelolaan produk.",
        "contact": {
            "email": "ahmadputrafirdaus@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/products": {
            "get": {
                "summary": "Daftar produk",
                "description": "Mengambil daftar semua produk atau mencari produk berdasarkan nama.",
                "operationId": "c1ff6f862214e1896c59cfe0491ce0e8",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nama produk untuk pencarian",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar produk berhasil diambil"
                    }
                }
            },
            "post": {
                "summary": "Tambah produk",
                "description": "Menyimpan produk baru ke database.",
                "operationId": "399770d33fa4672f1216707d0e236763",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Produk Baru"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Deskripsi produk"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produk berhasil disimpan"
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "summary": "Detail produk",
                "description": "Mengambil detail produk berdasarkan ID.",
                "operationId": "b46f4aa73232d0368eee698a082342a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID produk",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail produk berhasil diambil"
                    },
                    "404": {
                        "description": "Produk tidak ditemukan"
                    }
                }
            },
            "put": {
                "summary": "Update produk",
                "description": "Mengupdate data produk berdasarkan ID.",
                "operationId": "1419d479f9340e30e67fe21a2ca525b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID produk",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Produk Baru"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Deskripsi produk"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 12000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produk berhasil diupdate"
                    },
                    "404": {
                        "description": "Produk tidak ditemukan"
                    }
                }
            },
            "delete": {
                "summary": "Hapus produk",
                "description": "Menghapus produk berdasarkan ID.",
                "operationId": "fe34b364a3f192305598b65d04d06b5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID produk",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produk berhasil dihapus"
                    },
                    "404": {
                        "description": "Produk tidak ditemukan"
                    }
                }
            }
        }
    }
}